FAT2-371 moved isActive to shared base, usage in recruit and ass.
