+++ src/SFA.DAS.SharedOuterApi.UnitTests/InnerApi/Responses/StandardApiResponseBaseTests/TestStandardResponse.cs
@@ -1,0 +1,9 @@
+﻿using SFA.DAS.SharedOuterApi.InnerApi.Responses;
+
+namespace SFA.DAS.SharedOuterApi.UnitTests.InnerApi.Responses.StandardApiResponseBaseTests
+{
+    public class TestStandardResponse : StandardApiResponseBase
+    {
+            
+    }
+}
