+++ src/SFA.DAS.SharedOuterApi.UnitTests/InnerApi/Responses/StandardApiResponseBaseTests/WhenGettingMaxFunding.cs
@@ -5,15 +5,10 @@ using FluentAssertions;
 using NUnit.Framework;
 using SFA.DAS.SharedOuterApi.InnerApi.Responses;
 
-namespace SFA.DAS.SharedOuterApi.UnitTests.InnerApi.Responses
+namespace SFA.DAS.SharedOuterApi.UnitTests.InnerApi.Responses.StandardApiResponseBaseTests
 {
-    public class WhenMappingStandardApiResponse
+    public class WhenGettingMaxFunding
     {
-        private class TestStandardResponse : StandardApiResponseBase
-        {
-            
-        }
-        
         [Test]
         public void Then_If_There_Is_No_Available_Funding_Zero_Is_Returned_For_MaxFunding()
         {
@@ -26,20 +21,6 @@ namespace SFA.DAS.SharedOuterApi.UnitTests.InnerApi.Responses
             
             //Assert
             standard.MaxFunding.Should().Be(0);
-        }
-
-        [Test]
-        public void Then_If_There_Is_No_Available_Funding_Zero_Is_Returned_For_TypicalDuration()
-        {
-            //Arrange/Act
-            var standard =
-                new TestStandardResponse
-                {
-                    ApprenticeshipFunding = new List<ApprenticeshipFunding>()
-                };
-
-            //Assert
-            standard.TypicalDuration.Should().Be(0);
         }
 
         [Test, AutoData]
@@ -90,33 +71,7 @@ namespace SFA.DAS.SharedOuterApi.UnitTests.InnerApi.Responses
             standard.MaxFunding.Should().Be(fundingPrice);
         }
 
-        [Test, AutoData]
-        public void Then_The_Typical_Duration_Is_Used(int notDuration, int duration)
-        {
-            //Arrange / Act
-            var standard = new TestStandardResponse
-            {
-                ApprenticeshipFunding = new List<ApprenticeshipFunding>
-                {
-                    new ApprenticeshipFunding
-                    {
-                        EffectiveFrom = DateTime.UtcNow.AddDays(-10),
-                        EffectiveTo = DateTime.UtcNow.AddDays(-9),
-                        Duration = notDuration
-                    },
-                    new ApprenticeshipFunding
-                    {
-                        EffectiveFrom = DateTime.UtcNow.AddDays(-1),
-                        EffectiveTo = null,
-                        Duration = duration
-                    }
-                }
-
-            };
-
-            //Assert
-            standard.TypicalDuration.Should().Be(duration);
-        }
+        
 
         [Test, AutoData]
         public void Then_The_Future_Price_Is_Not_Used(int fundingPrice, int notFundingPrice)
